Remove dist folder from tracking
